Moved "Open in new tab" toggle to separate component, using the new API
